Add child character and enemy classes
