+++ Aura/Source/Aura/Private/Character/AuraCharacter.cpp
@@ -1,0 +1,5 @@
+// Copyright Mattman
+
+
+#include "Character/AuraCharacter.h"
+
