+++ Aura/Source/Aura/Private/Character/AuraCharacterBase.cpp
@@ -1,0 +1,15 @@
+// Copyright Mattman
+
+
+#include "Character/AuraCharacterBase.h"
+
+// Sets default values
+AAuraCharacterBase::AAuraCharacterBase()
+{
+	PrimaryActorTick.bCanEverTick = false;
+}
+
+void AAuraCharacterBase::BeginPlay()
+{
+	Super::BeginPlay();
+}
