+++ Aura/Source/Aura/Private/Character/AuraEnemy.cpp
@@ -1,0 +1,5 @@
+// Copyright Mattman
+
+
+#include "Character/AuraEnemy.h"
+
