--- Aura/Source/Aura/Public/AuraCharacterBase.h
@@ -1,29 +1,0 @@
-// Copyright Mattman
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "GameFramework/Character.h"
-#include "AuraCharacterBase.generated.h"
-
-UCLASS()
-class AURA_API AAuraCharacterBase : public ACharacter
-{
-	GENERATED_BODY()
-
-public:
-	// Sets default values for this character's properties
-	AAuraCharacterBase();
-
-protected:
-	// Called when the game starts or when spawned
-	virtual void BeginPlay() override;
-
-public:	
-	// Called every frame
-	virtual void Tick(float DeltaTime) override;
-
-	// Called to bind functionality to input
-	virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;
-
-};
