+++ Aura/Source/Aura/Public/Character/AuraCharacter.h
@@ -1,0 +1,17 @@
+// Copyright Mattman
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Character/AuraCharacterBase.h"
+#include "AuraCharacter.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class AURA_API AAuraCharacter : public AAuraCharacterBase
+{
+	GENERATED_BODY()
+	
+};
