+++ Aura/Source/Aura/Public/Character/AuraCharacterBase.h
@@ -1,0 +1,19 @@
+// Copyright Mattman
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/Character.h"
+#include "AuraCharacterBase.generated.h"
+
+UCLASS(Abstract)
+class AURA_API AAuraCharacterBase : public ACharacter
+{
+	GENERATED_BODY()
+
+public:
+	AAuraCharacterBase();
+
+protected:
+	virtual void BeginPlay() override;
+};
