+++ Aura/Source/Aura/Public/Character/AuraEnemy.h
@@ -1,0 +1,17 @@
+// Copyright Mattman
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Character/AuraCharacterBase.h"
+#include "AuraEnemy.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class AURA_API AAuraEnemy : public AAuraCharacterBase
+{
+	GENERATED_BODY()
+	
+};
